Fix regression in `gpio::v1` compatibility
At reset, the `gpio::v1` module gives `Pin`s to users in `FloatingInput`
mode, but this is not correct, because the pins are actually in
`FloatingDisabled` mode. This problem was never discovered, because
users always called a function to change the pin mode before using a
given pin, even when that function was `.into_floating_input()`.

In a recent change to `gpio::v2`, the `into_mode` function was altered
to only update the register settings when actually changing modes. The
goal of this change was to prevent spurious writes to the GPIO
registers.

Combined with the incorrect behavior of `gpio::v1`, this change created
a situation where it was impossible to use a pin as a floating input
without first converting it to some other mode and back. Pins would be
labelled as `FloatingInput`, even though they were actually
`FloatingDisabled`. But calls to `.into_floating_input()` would have no
effect, because there was no change in `PinMode`.

Introduce an `into_mode` method for `v1::Pin`s that mirrors the design
of the corresponding method of `v2::Pin`, but in this case change the
register values unconditionally. This does not fix the fundamental flaw
in `gpio::v1`, but it does return it to the pre-existing behavior.